Add runtime prerequisites for Lua and fix assembler escape sequences
Runtime: implement setjmp/longjmp, add clock()/strftime() time stubs,
fix difftime() to return double, expand float.h/locale.h/math.h/stdio.h/
string.h/time.h headers, add log2(), freopen(), ungetc(), setvbuf(),
tmpfile(), strcoll(), strerror() stubs, and signal.h stub.

Assembler: add \a \b \f \v escape sequences to .ascii directive parser.
The missing \b and \f caused LLVM-emitted character tables to be silently
corrupted (e.g. 0x08 stored as 'b'=0x62), breaking Lua's lexer ctype table.

+++ runtime/time_extra.c
@@ -149,6 +149,6 @@ char *ctime(const time_t *timer) {
     return asctime(localtime(timer));
 }
 
-long long difftime(time_t time1, time_t time0) {
-    return (long long)time1 - (long long)time0;
+double difftime(time_t time1, time_t time0) {
+    return (double)((long long)time1 - (long long)time0);
 }
